+++ sdk/eventhub/azure-eventhub/tests/unittest/test_client_creation.py
@@ -1,4 +1,11 @@
-from azure.eventhub import EventHubProducerClient, EventHubConsumerClient, TransportType
+#-------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for
+# license information.
+#--------------------------------------------------------------------------
+
+import time
+from azure.eventhub import EventHubProducerClient, EventHubConsumerClient, TransportType, RetryMode
 
 
 def test_custom_endpoint():
@@ -123,3 +130,19 @@ def test_custom_certificate():
         connection_verify='D:/local/certfile'
     )
     assert consumer._config.connection_verify == 'D:/local/certfile'
+
+def test_backoff_fixed_retry():
+    client = EventHubProducerClient(
+        'fake.host.com',
+        'fake_eh',
+        None,
+        retry_mode=RetryMode.FIXED
+    )
+    backoff = client._config.backoff_factor
+    start_time = time.time()
+    client._backoff(retried_times=1, last_exception=Exception('fake'), timeout_time=None)
+    sleep_time = time.time() - start_time
+    # exp = 0.8 * (2 ** 1) = 1.6
+    # time.sleep() in _backoff will take AT LEAST time 'exp' for RetryMode.EXPONENTIAL
+    # check that fixed is less than 'exp'
+    assert sleep_time < backoff * (2 ** 1)